+++ packages/eslint-plugin/src/rules/space-before-function-paren.ts
@@ -139,7 +139,8 @@ export default util.createRule<Options, MessageIds>({
         return;
       }
 
-      let leftToken: TSESTree.Token, rightToken: TSESTree.Token;
+      let leftToken: TSESTree.Token;
+      let rightToken: TSESTree.Token;
       if (node.typeParameters) {
         leftToken = sourceCode.getLastToken(node.typeParameters)!;
         rightToken = sourceCode.getTokenAfter(leftToken)!;